Move TextTaskInput to shared components
