+++ ground/src/main/java/com/google/android/ground/ui/datacollection/components/TextTaskInput.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.android.ground.ui.datacollection.tasks.text
+package com.google.android.ground.ui.datacollection.components
 
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.wrapContentHeight
@@ -24,14 +24,11 @@ import androidx.compose.material3.TextField
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.google.android.ground.ExcludeFromJacocoGeneratedReport
 import com.google.android.ground.ui.theme.AppTheme
-
-const val INPUT_TEXT_TEST_TAG: String = "text task input test tag"
 
 @Composable
 fun TextTaskInput(
@@ -48,8 +45,7 @@ fun TextTaskInput(
         .wrapContentHeight(align = Alignment.Top)
         // TODO: Add horizontal padding as 16.dp when global padding is removed.
         // Issue URL: https://github.com/google/ground-android/issues/2976
-        .padding(vertical = 8.dp)
-        .testTag(INPUT_TEXT_TEST_TAG),
+        .padding(vertical = 8.dp),
     textStyle = MaterialTheme.typography.bodyLarge,
     singleLine = true,
     keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Text),
